% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_doc_ENSEMBL_Reference.R
\docType{data}
\name{GRCm38_ref}
\alias{GRCm38_ref}
\title{Information on 55,435 genes with ENSEMBL ID (build GRCm38/mm10).}
\format{
A list with 3 elements:
\describe{
\item{GRCm38G}{dataframe, information on genes}
\item{chr_len_M}{named integer vector, information on chromosomes}
\item{GRCm38T}{dataframe, information on transcripts}
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
GRCm38_ref
}
\description{
A reference dataset containing the information on 55,435 genes with ENSEMBL ID (build GRCm38/mm10) for vigilante.knights (the main package).
}
\keyword{datasets}
