% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vks_doc_HKG_Reference.R
\docType{data}
\name{hkg_reference}
\alias{hkg_reference}
\title{Information on 3,598 housekeeping genes.}
\format{
A dataframe with 3 columns:
\describe{
\item{ENSG}{character, ENSEMBL ID}
\item{Gene}{character, gene name}
\item{Pathway}{character, related pathogenic pathway (if applicable)}
}
}
\source{
\url{https://doi.org/10.1016/j.tig.2013.05.010}
}
\usage{
hkg_reference
}
\description{
A reference dataset containing the information on 3,598 housekeeping genes for vigilante.knights (the main package).
}
\keyword{datasets}
