% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vks_doc_Oncogene_Reference.R
\docType{data}
\name{oncogene_reference}
\alias{oncogene_reference}
\title{Information on 2,370 oncogenes.}
\format{
A dataframe with 3 columns:
\describe{
\item{ENSG}{character, ENSEMBL ID}
\item{Gene}{character, gene name}
\item{Pathway}{character, related pathogenic pathway (if applicable)}
}
}
\source{
\url{https://doi.org/10.1186/s13059-018-1612-0}
}
\usage{
oncogene_reference
}
\description{
A reference dataset containing the information on 2,370 oncogenes for vigilante.knights (the main package).
}
\keyword{datasets}
