% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vks_doc_CHM_T_Example.R
\docType{data}
\name{transcript_commonName}
\alias{transcript_commonName}
\title{Example input dataset for module: CHM_T in vigilante.knights.}
\format{
A dataframe with 2 columns:
\describe{
\item{Transcript}{character, transcript name}
\item{Common_Name}{character, more commonly used or known name of the corresponding transcript, may contain non-ASCII characters (already converted to UTF-8 encoding)}
}
}
\source{
\url{http://www.ensembl.org/} + manually curated
}
\usage{
transcript_commonName
}
\description{
An example input dataset for module: CHM_T in vigilante.knights (the main package), containing the information on 24 transcripts.
}
\keyword{datasets}
