% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vks_doc_CHM_T_Example.R
\docType{data}
\name{transcript_selection}
\alias{transcript_selection}
\title{Example input dataset for module: CHM_T in vigilante.knights.}
\format{
A dataframe with 3 columns:
\describe{
\item{ENST}{character, ENSEMBL transcript ID}
\item{ENSG}{character, ENSEMBL gene ID}
\item{Gene}{character, gene name}
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
transcript_selection
}
\description{
An example input dataset for module: CHM_T in vigilante.knights (the main package), containing the information on 24 transcripts.
}
\keyword{datasets}
