% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_doc_ENSEMBL_Reference.R
\docType{data}
\name{GRCh37_ref}
\alias{GRCh37_ref}
\title{Information of 33,579 genes with ENSEMBL ID (build GRCh37/hg19).}
\format{
A list with 3 elements:
\describe{
\item{GRCh37G}{dataframe, information of genes}
\item{chr_len_H}{named integer vector, information of chromosomes}
\item{GRCh37T}{dataframe, information of transcripts}
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
GRCh37_ref
}
\description{
A reference dataset containing the information of 33,579 genes with ENSEMBL ID (build GRCh37/hg19) for vigilante.knights (the main package).
}
\keyword{datasets}
