% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_doc_ENSEMBL_Reference.R
\docType{data}
\name{GRCh38_ref}
\alias{GRCh38_ref}
\title{Information of 38,965 genes with ENSEMBL ID (build GRCh38/hg38).}
\format{
A list with 3 elements:
\describe{
\item{GRCh38G}{dataframe, information of genes}
\item{chr_len_H}{named integer vector, information of chromosomes}
\item{GRCh38T}{dataframe, information of transcripts}
}
}
\source{
\url{http://www.ensembl.org/}
}
\usage{
GRCh38_ref
}
\description{
A reference dataset containing the information of 38,965 genes with ENSEMBL ID (build GRCh38/hg38) for vigilante.knights (the main package).
}
\keyword{datasets}
